Outline: refactor outline view
