+++ src/devtools/client/debugger/src/components/Outline/Outline.css
